made a major feature in code
Details: i made a new way to access css by including the css in a include folder, then include a file from that folder to the assigned css. I added a email feature for testing. i changed all the urlroot links in all the buttons that were necessary. i added a try catch in the delete function. i added a html_entity_decode funtion for the continent so i dont repeat it in the update select option, so if the value from database is a match to one of those options display that value.